Added drop down Menu for frequencies

+++ app/src/main/java/com/track_it/persistence/hsqldb/SubscriptionPersistenceHSQLDB.java
@@ -5,6 +5,7 @@ import android.util.Log;
 
 import com.track_it.domainobject.SubscriptionObj;
 import com.track_it.logic.exception.DataBaseException;
+import com.track_it.logic.exception.DataBaseSubNotFoundException;
 import com.track_it.persistence.SubscriptionPersistence;
 
 import java.sql.Connection;
@@ -70,8 +71,16 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
         return AllSubscriptions;
     }
 
-    public void editSubscriptionByID(int subscriptionID, SubscriptionObj newSubscriptionDetails)
+    public void editSubscriptionByID(int subscriptionIDToEdit, SubscriptionObj newSubscriptionDetails) throws DataBaseException
     {
+
+        //If a subscription with id of subscriptionID is not found in database throw a DataBaseException exception
+        if ( !subscriptionInDatabase(subscriptionIDToEdit) )
+        {
+            throw new DataBaseSubNotFoundException("");
+        }
+
+
         SubscriptionObj subToReturn= null;
 
         try (Connection connection = connect())
@@ -81,7 +90,7 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
             statement.setString(1, newSubscriptionDetails.getName());
             statement.setInt(2, newSubscriptionDetails.getTotalPaymentInCents());
             statement.setString(3, newSubscriptionDetails.getPaymentFrequency());
-            statement.setInt(4, subscriptionID);
+            statement.setInt(4, subscriptionIDToEdit);
 
            statement.executeUpdate();
 
@@ -95,7 +104,7 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
 
     }
 
-    public void addSubscriptionToDB(SubscriptionObj subscriptionToAdd)
+    public void addSubscriptionToDB(SubscriptionObj subscriptionToAdd) throws DataBaseException
     {
 
         try (Connection connection = connect()) {
@@ -122,11 +131,19 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
 
     }
 
-    public void removeSubscriptionByID(int subscriptionToRemove)
+    public void removeSubscriptionByID(int subscriptionIDToRemove) throws DataBaseException
     {
+        //If a subscription with id of subscriptionID is not found in database throw a DataBaseException exception
+        if ( !subscriptionInDatabase(subscriptionIDToRemove) )
+        {
+            throw new DataBaseSubNotFoundException();
+        }
+
+
+        // Otherwise remove the subscription from the database
         try (Connection connection = connect()) {
             final PreparedStatement statement = connection.prepareStatement("DELETE FROM SUBSCRIPTIONS WHERE ID = ?");
-              statement.setInt(1, subscriptionToRemove);
+              statement.setInt(1, subscriptionIDToRemove);
 
             statement.executeUpdate(); // execute delete statement
             statement.close();
@@ -140,14 +157,20 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
         }
     }
 
-    public  SubscriptionObj getSubscriptionByID(int subscriptionID)
+    public  SubscriptionObj getSubscriptionByID(int subscriptionIDtoGet) throws DataBaseException
     {
+        //If a subscription with id of subscriptionID is not found in database throw a DataBaseException exception
+        if ( !subscriptionInDatabase(subscriptionIDtoGet) )
+        {
+            throw new DataBaseSubNotFoundException();
+        }
 
-        SubscriptionObj subToReturn= null;
+        // Otherwise get the subscription from the database, and return it
+        SubscriptionObj subToReturn = null;
         try (Connection connection = connect())
         {
             final PreparedStatement statement = connection.prepareStatement("SELECT * FROM SUBSCRIPTIONS WHERE id = ?");
-            statement.setInt(1, subscriptionID);
+            statement.setInt(1, subscriptionIDtoGet);
 
             final ResultSet resultSet = statement.executeQuery();
             if (resultSet.next())
@@ -165,9 +188,35 @@ public class SubscriptionPersistenceHSQLDB implements SubscriptionPersistence {
             throw new DataBaseException(e.getMessage());
 
         }
-
     }
 
+
+    //Returns true if a subscription with id of subscriptionID is found in database, else return false
+    private boolean subscriptionInDatabase(int subscriptionID)  throws DataBaseException
+    {
+
+        boolean found = false;
+
+        try (Connection connection = connect())
+        {
+            final PreparedStatement statement = connection.prepareStatement("SELECT * FROM SUBSCRIPTIONS WHERE id = ?");
+            statement.setInt(1, subscriptionID);
+
+            final ResultSet resultSet = statement.executeQuery();
+            if (resultSet.next())
+            {
+
+                found = true;
+            }
+            return found;
+        }
+
+        catch (final SQLException e) {
+            Log.e("Connect SQL", e.getMessage() + e.getSQLState());
+            e.printStackTrace();
+            throw new DataBaseException(e.getMessage());
+        }
+    }
 
 
 }
